refactor: 인증 관련 클래스 auth 패키지로 분리
