+++ src/main/java/com/ssafyss/board_practice/domain/auth/api/AuthController.java
@@ -1,11 +1,11 @@
-package com.ssafyss.board_practice.domain.user.api;
+package com.ssafyss.board_practice.domain.auth.api;
 
-import com.ssafyss.board_practice.domain.user.application.UserService;
-import com.ssafyss.board_practice.domain.user.constants.SuccessMessages;
-import com.ssafyss.board_practice.domain.user.dto.CheckEmailRequest;
-import com.ssafyss.board_practice.domain.user.dto.CreateUserRequest;
-import com.ssafyss.board_practice.domain.user.dto.SignInRequest;
-import com.ssafyss.board_practice.domain.user.dto.SignInResponse;
+import com.ssafyss.board_practice.domain.auth.application.AuthService;
+import com.ssafyss.board_practice.domain.auth.constants.SuccessMessages;
+import com.ssafyss.board_practice.domain.auth.dto.CheckEmailRequest;
+import com.ssafyss.board_practice.domain.auth.dto.CreateUserRequest;
+import com.ssafyss.board_practice.domain.auth.dto.SignInRequest;
+import com.ssafyss.board_practice.domain.auth.dto.SignInResponse;
 import com.ssafyss.board_practice.global.dto.BaseResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,17 +14,17 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/user")
-public class UserController {
-    private final UserService userService;
+@RequestMapping("/auth")
+public class AuthController {
+    private final AuthService authService;
 
-    public UserController(UserService userService) {
-        this.userService = userService;
+    public AuthController(AuthService authService) {
+        this.authService = authService;
     }
 
     @PostMapping("/checkEmail")
     public ResponseEntity<BaseResponse> checkEmail(@RequestBody CheckEmailRequest request) {
-        userService.checkEmail(request.getEmail());
+        authService.checkEmail(request.getEmail());
         return ResponseEntity.ok(new BaseResponse.Builder()
                 .message(SuccessMessages.EMAIL_AVAILABLE)
                 .build());
@@ -32,7 +32,7 @@ public class UserController {
 
     @PostMapping("/signUp")
     public ResponseEntity<BaseResponse> signUp(@RequestBody CreateUserRequest request) {
-        userService.signUp(request.getEmail(), request.getPassword());
+        authService.signUp(request.getEmail(), request.getPassword());
         return ResponseEntity.ok(new BaseResponse.Builder()
                 .message(SuccessMessages.SIGN_UP_SUCCESS)
                 .build());
@@ -40,7 +40,7 @@ public class UserController {
 
     @PostMapping("/signIn")
     public ResponseEntity<SignInResponse> signIn(@RequestBody SignInRequest request) {
-        SignInResponse signInResponse = userService.signIn(request.getEmail(), request.getPassword());
+        SignInResponse signInResponse = authService.signIn(request.getEmail(), request.getPassword());
         return ResponseEntity.ok(signInResponse);
     }
