+++ src/main/java/com/ssafyss/board_practice/domain/auth/constants/ErrorMessages.java
@@ -1,4 +1,4 @@
-package com.ssafyss.board_practice.domain.user.constants;
+package com.ssafyss.board_practice.domain.auth.constants;
 
 public class ErrorMessages {
     public static final String EMAIL_DUPLICATED = "중복된 이메일입니다";
