+++ src/main/java/com/ssafyss/board_practice/domain/auth/constants/SuccessMessages.java
@@ -1,4 +1,4 @@
-package com.ssafyss.board_practice.domain.user.constants;
+package com.ssafyss.board_practice.domain.auth.constants;
 
 public class SuccessMessages {
     public static final String EMAIL_AVAILABLE = "사용할 수 있는 이메일입니다";
