+++ src/main/java/com/ssafyss/board_practice/domain/auth/dto/CheckEmailRequest.java
@@ -1,10 +1,10 @@
-package com.ssafyss.board_practice.domain.user.dto;
+package com.ssafyss.board_practice.domain.auth.dto;
 
 public class CheckEmailRequest {
     private String email;
 
     public CheckEmailRequest() {
-        
+
     }
 
     public String getEmail() {
