+++ src/main/java/com/ssafyss/board_practice/domain/auth/exception/SignInFailedException.java
@@ -1,4 +1,4 @@
-package com.ssafyss.board_practice.domain.user.exception;
+package com.ssafyss.board_practice.domain.auth.exception;
 
 public class SignInFailedException extends RuntimeException {
     public SignInFailedException(String message) {
